minor updates background notifications

+++ Handshake/iphone/controllers/editor/AddressEditController.m
@@ -92,7 +92,7 @@
     [tableView cellForRowAtIndexPath:[NSIndexPath indexPathForRow:self.selectedLabel + 5 inSection:0]].accessoryType = UITableViewCellAccessoryNone;
     [tableView cellForRowAtIndexPath:indexPath].accessoryType = UITableViewCellAccessoryCheckmark;
     
-    self.selectedLabel = indexPath.row - 5;
+    self.selectedLabel = (int)indexPath.row - 5;
     
     [self updateSaveButton];
 }
